requirements.txt
requests
packageurl-python